separate WeihanLi.Core project
