+++ WeihanLi.Common.slnx
@@ -7,6 +7,7 @@
     <Project Path="samples/DotNetCoreSample/DotNetCoreSample.csproj" />
   </Folder>
   <Folder Name="/src/">
+    <Project Path="src/WeihanLi.Core/WeihanLi.Core.csproj" />
     <Project Path="src/WeihanLi.Common.Logging.Serilog/WeihanLi.Common.Logging.Serilog.csproj" />
     <Project Path="src/WeihanLi.Common/WeihanLi.Common.csproj" />
     <Project Path="src/WeihanLi.Extensions.Hosting/WeihanLi.Extensions.Hosting.csproj" />
