+++ src/WeihanLi.Common/README.md
@@ -2,7 +2,9 @@
 
 ## Intro
 
-`WeihanLi.Common` is a common library based on .NETStandard2.0
+`WeihanLi.Common` is a common library based on .NETStandard2.0.
+
+Low-dependency, AOT-compatible extensions and primitives now live in `WeihanLi.Core`. `WeihanLi.Common` references `WeihanLi.Core` and keeps the heavier DI, AOP, logging, data access, HTTP, JSON, and configuration integrations.
 
 ## Features
