+++ src/WeihanLi.Core/Extensions/CollectionExtension.cs
