+++ src/WeihanLi.Core/Extensions/CompressionExtension.cs
