+++ src/WeihanLi.Core/Extensions/EnumerableExtension.cs
