+++ src/WeihanLi.Core/Extensions/ExceptionExtension.cs
