+++ src/WeihanLi.Core/Extensions/FuncExtension.cs
