+++ src/WeihanLi.Core/Extensions/IOExtension.cs
