+++ src/WeihanLi.Core/Extensions/TaskExtension.cs
