+++ src/WeihanLi.Core/Guard.cs
