+++ src/WeihanLi.Core/Helpers/Combinatorics/GenerateOption.cs
