+++ src/WeihanLi.Core/Helpers/Combinatorics/SmallPrimeUtility.cs
