+++ src/WeihanLi.Core/Helpers/Combinatorics/Variations.cs
