+++ src/WeihanLi.Core/Helpers/HashHelper.cs
