+++ src/WeihanLi.Core/Helpers/SecurityHelper.cs
