+++ src/WeihanLi.Core/Models/DataOperationType.cs
