+++ src/WeihanLi.Core/Models/KeyEntry.cs
