+++ src/WeihanLi.Core/Models/Result.cs
