+++ src/WeihanLi.Core/Models/ResultStatus.cs
