+++ src/WeihanLi.Core/Models/TenantInfo.cs
