+++ src/WeihanLi.Core/Models/ValidationResult.cs
