+++ src/WeihanLi.Core/README.md
@@ -1,0 +1,5 @@
+# WeihanLi.Core
+
+AOT-compatible core helpers, extensions, models, and utility primitives split from `WeihanLi.Common`.
+
+This package intentionally avoids dependencies on the heavier `WeihanLi.Common` areas such as dependency injection, AOP, logging, data access, HTTP requester helpers, JSON integration, and configuration integration.
